+++ R/add_HeadTail.R
@@ -5,8 +5,8 @@
 #' (image date) column which are (\code{Date} variable).
 #' @inheritParams check_input
 #' @inheritParams season
-#' 
-#' @param trs If nmissing < trs*nptperyear (little missing), this year is 
+#'
+#' @param trs If nmissing < trs*nptperyear (little missing), this year is
 #' include to extract phenology; if \code{FALSE}, this year is excluded.
 #'
 #' @note \code{date} is image date; \code{t} is compositing date.
@@ -17,17 +17,17 @@
 #' @examples
 #' library(phenofit)
 #' data("MOD13A1")
-#' 
+#'
 #' dt <- tidy_MOD13.gee(MOD13A1$dt)
 #' st <- MOD13A1$st
-#' 
+#'
 #' sitename <- dt$site[1]
 #' d     <- dt[site == sitename, ] # get the first site data
 #' sp    <- st[site == sitename, ] # station point
-#' 
+#'
 #' nptperyear = 23
 #' dnew     <- add_HeadTail(d, nptperyear = nptperyear) # add one year in head and tail
-#' 
+#'
 #' @export
 add_HeadTail <- function(d, south = FALSE, nptperyear, trs = 0.45){
     if (missing(nptperyear)){